2024-02-08 수업내용+ test

+++ src/main/java/com/ohgiraffers/section01/conditional/test/Level02_2.java
@@ -12,7 +12,7 @@ public class Level02_2 {
         int math = sc.nextInt();
         System.out.print("영어 점수를 입력하세요 : ");
         int english = sc.nextInt();
-        int average = (math+english)/2;
+        double average = (math+english)/2;
         if ( (average >= 60) && (math >= 40) && (english >=40) ) {
 
             System.out.println("합격");
@@ -22,3 +22,26 @@ public class Level02_2 {
     }
 
 }
+// 또 다른 풀이
+//        Scanner sc = new Scanner(System.in);
+//        System.out.print("수학 점수를 입력해주세요 : ");
+//        int math = sc.nextInt();
+//
+//        System.out.print("영어 점수를 입력해주세요 : ");
+//        int english = sc.nextInt();
+//        double avg = (math + english) / 2;
+//
+//        if(avg < 60) {
+//
+//            System.out.println("시험 불합격");
+//        } else if (english < 40 || math < 40){
+//
+//            System.out.println("한 과목 과락으로 인한 시험 불합격");
+//        } else {
+//            System.out.println("시험에 합격했습니다.");
+//        }
+
+
+
+
+
